feat: Cluster -> Environment - namespace labels support, code refactor
