--- src/Pages/GlobalConfigurations/ClustersAndEnvironments/EnvironmentLabels/index.ts
@@ -1,1 +1,0 @@
-export * from './EnvironmentLabels'
