--- src/Pages/GlobalConfigurations/ClustersAndEnvironments/clustersAndEnvironments.types.ts
@@ -1,10 +1,0 @@
-export interface ClusterNamespacesLabel {
-    key: string
-    value: string
-}
-
-export interface ClusterNamespacesDTO {
-    name: string
-    labels: ClusterNamespacesLabel[]
-    resourceVersion: string
-}
